+++ src/backend/actions/edit-action.ts
@@ -1,5 +1,4 @@
-import Action from './action.interface'
-import RecordJSON from '../decorators/record-json.interface'
+import Action, { RecordActionResponse } from './action.interface'
 import NotFoundError from '../utils/not-found-error'
 
 /**
@@ -11,7 +10,7 @@ import NotFoundError from '../utils/not-found-error'
  *
  * Uses {@link EditAction} component to render form
  */
-const EditAction: Action = {
+const EditAction: Action<RecordActionResponse> = {
   name: 'edit',
   isVisible: true,
   actionType: 'record',
@@ -22,11 +21,11 @@ const EditAction: Action = {
    *
    * To invoke this action use {@link ApiClient#recordAction}
    *
-   * @return  {EditActionResponse}  populated record
+   * @return  {RecordActionResponse}  populated record
    * @implements Action.handler
    * @memberof module:EditAction
    */
-  handler: async (request, response, data): Promise<EditActionResponse> => {
+  handler: async (request, response, data) => {
     const { record } = data
     if (!record) {
       throw new NotFoundError([
@@ -42,28 +41,21 @@ const EditAction: Action = {
         redirectUrl: data.h.recordActionUrl({
           resourceId: data.resource.id(), recordId: record.id(), actionName: 'show',
         }),
+        notice: {
+          message: 'Successfully updated the record',
+          type: 'success',
+        },
         record: record.toJSON(data.currentAdmin),
       }
     }
-    return { record: record.toJSON(data.currentAdmin) }
+    return {
+      record: record.toJSON(data.currentAdmin),
+      notice: {
+        message: 'There are validation errors - check them out below.',
+        type: 'error',
+      },
+    }
   },
 }
 
 export default EditAction
-
-/**
- * Type of response returned by {@link module:EditAction}
- * @memberof module:EditAction
- */
-export type EditActionResponse = {
-  /**
-   * in case of success it fills this filed
-   * to indicate that there should be
-   * redirect after the action.
-   */
-  redirectUrl?: string;
-  /**
-   * Updated record
-   */
-  record: RecordJSON;
-}