+++ src/backend/utils/validation-error.ts
@@ -1,23 +1,52 @@
-import { RecordError } from '../adapters/base-record'
+
+/**
+ * Record Error
+ * @memberof ValidationError
+ */
+export type RecordError = {
+  /**
+   * error type (i.e. required)
+   */
+  type: string;
+  /**
+   * human readible message
+   */
+  message: string;
+}
+
+/**
+ * Property Errors
+ * @memberof ValidationError
+ */
+export type PropertyErrors = {
+  [key: string]: RecordError;
+}
 
 /**
  * Error which is thrown when there are validation errors with records
  * @category Errors
  */
 class ValidationError extends Error {
-  public errors: {[key: string]: RecordError}
+  /**
+   * Validation errors for all properties
+   */
+  public propertyErrors: PropertyErrors
 
   /**
-   * @param  {String} message   custom message
-   * @param  {Object} errors    error messages
-   * @param  {String} errors.{...}    error for particular field where ... is a
-   *                                  {@link BaseProperty#path}
-   * @param  {String} errors.{...}.message   human readible message
-   * @param  {String} errors.{...}.type      string type (i.e. required)
+   * One root validatin error i.e. thrown when user wants to perform
+   * an action which violates foreign key constraint
    */
-  constructor(message: string, errors: {[key: string]: RecordError}) {
+  public baseError: RecordError | null
+
+  /**
+   * @param {string} message   custom message
+   * @param {PropertyErrors} propertyErrors     error messages
+   * @param {RecordError} [baseError]           base error
+   */
+  constructor(message: string, propertyErrors: PropertyErrors, baseError?: RecordError) {
     super(message)
-    this.errors = errors
+    this.propertyErrors = propertyErrors
+    this.baseError = baseError || null
     this.message = message || 'Resource cannot be stored because of validation errors'
     this.name = 'ValidationError'
   }