+++ src/frontend/components/actions/new.tsx
@@ -78,16 +78,12 @@ class New extends React.Component<ActionProps & AddNoticeProps & RouteComponentP
       payload: formData,
       headers: { 'Content-Type': 'multipart/form-data' },
     }).then((response) => {
+      if (response.data.notice) {
+        addNotice(response.data.notice)
+      }
       if (response.data.redirectUrl) {
-        addNotice({
-          message: 'Record has been successfully created!',
-        })
         history.push(response.data.redirectUrl)
       } else {
-        addNotice({
-          type: 'error',
-          message: 'There were errors in the record object. Check them out',
-        })
         this.setState(state => ({
           record: {
             ...state.record,